Add navigation enums and UserRouter for user operations
Introduced several enumeration classes to enable navigation and user related operations, including functionality for logging in or out, creating and deleting accounts, etc. These changes simplify the interactions in the main application and improve the structure and readability of the code. Additionally, replaced parser values with enums for clearer function execution.